+++ app/src/main/java/woowacourse/shopping/presentation/cart/CartActivity.kt
@@ -1,0 +1,24 @@
+package woowacourse.shopping.presentation.cart
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import woowacourse.shopping.presentation.cart.screen.CartScreen
+import woowacourse.shopping.presentation.theme.androidshoppingTheme
+import kotlin.uuid.ExperimentalUuidApi
+
+class CartActivity : ComponentActivity() {
+    @OptIn(ExperimentalUuidApi::class)
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+
+        @OptIn(ExperimentalUuidApi::class)
+        setContent {
+            androidshoppingTheme {
+                CartScreen(onBack = { finish() })
+            }
+        }
+    }
+}
