+++ app/src/main/java/woowacourse/shopping/presentation/navigation/IntentKeys.kt
@@ -1,0 +1,5 @@
+package woowacourse.shopping.presentation.navigation
+
+object IntentKeys {
+    const val PRODUCT_ID = "woowacourse.shopping.product_id"
+}
