+++ app/src/main/java/woowacourse/shopping/presentation/productdetail/ProductDetailActivity.kt
@@ -1,0 +1,45 @@
+package woowacourse.shopping.presentation.productdetail
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import woowacourse.shopping.data.repository.ProductRepositoryImpl
+import woowacourse.shopping.presentation.navigation.IntentKeys
+import woowacourse.shopping.presentation.productdetail.screen.ProductDetailErrorScreen
+import woowacourse.shopping.presentation.productdetail.screen.ProductDetailScreen
+import woowacourse.shopping.presentation.theme.androidshoppingTheme
+import kotlin.uuid.ExperimentalUuidApi
+import kotlin.uuid.Uuid
+
+class ProductDetailActivity : ComponentActivity() {
+    @OptIn(ExperimentalUuidApi::class)
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+
+        val product =
+            intent
+                .getStringExtra(IntentKeys.PRODUCT_ID)
+                ?.toUuidOrNull()
+                ?.let { productId ->
+                    ProductRepositoryImpl.findProductById(productId)
+                }
+
+        setContent {
+            androidshoppingTheme {
+                if (product != null) {
+                    ProductDetailScreen(
+                        product = product,
+                        onClose = { finish() },
+                    )
+                } else {
+                    ProductDetailErrorScreen(onClose = { finish() })
+                }
+            }
+        }
+    }
+}
+
+@OptIn(ExperimentalUuidApi::class)
+private fun String.toUuidOrNull(): Uuid? = runCatching { Uuid.parse(this) }.getOrNull()
