--- app/src/main/java/woowacourse/shopping/ui/cart/CartActivity.kt
@@ -1,24 +1,0 @@
-package woowacourse.shopping.ui.cart
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import woowacourse.shopping.ui.cart.screen.CartScreen
-import woowacourse.shopping.ui.theme.androidshoppingTheme
-import kotlin.uuid.ExperimentalUuidApi
-
-class CartActivity : ComponentActivity() {
-    @OptIn(ExperimentalUuidApi::class)
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-
-        @OptIn(ExperimentalUuidApi::class)
-        setContent {
-            androidshoppingTheme {
-                CartScreen(onBack = { finish() })
-            }
-        }
-    }
-}
