--- app/src/main/java/woowacourse/shopping/ui/navigation/IntentKeys.kt
@@ -1,5 +1,0 @@
-package woowacourse.shopping.ui.navigation
-
-object IntentKeys {
-    const val PRODUCT_ID = "woowacourse.shopping.product_id"
-}
