--- app/src/main/java/woowacourse/shopping/ui/shopping/ProductListActivity.kt
@@ -1,39 +1,0 @@
-package woowacourse.shopping.ui.shopping
-
-import android.content.Intent
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import woowacourse.shopping.data.ProductFixture
-import woowacourse.shopping.domain.model.product.Products
-import woowacourse.shopping.ui.cart.CartActivity
-import woowacourse.shopping.ui.navigation.IntentKeys
-import woowacourse.shopping.ui.productdetail.ProductDetailActivity
-import woowacourse.shopping.ui.shopping.screen.ProductListScreen
-import woowacourse.shopping.ui.theme.androidshoppingTheme
-
-class ProductListActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        setContent {
-            androidshoppingTheme {
-                ProductListScreen(
-                    products = Products(ProductFixture.productList),
-                    onCartIconClick = {
-                        val intent = Intent(this, CartActivity::class.java)
-                        startActivity(intent)
-                    },
-                    onItemClick = { productId ->
-                        val intent =
-                            Intent(this, ProductDetailActivity::class.java).apply {
-                                putExtra(IntentKeys.PRODUCT_ID, productId)
-                            }
-                        startActivity(intent)
-                    },
-                )
-            }
-        }
-    }
-}
