--- app/src/main/java/woowacourse/shopping/ui/theme/Color.kt
@@ -1,19 +1,0 @@
-package woowacourse.shopping.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-val Purple80 = Color(0xFFD0BCFF)
-val PurpleGrey80 = Color(0xFFCCC2DC)
-val Pink80 = Color(0xFFEFB8C8)
-
-val Purple40 = Color(0xFF6650a4)
-val PurpleGrey40 = Color(0xFF625b71)
-val Pink40 = Color(0xFF7D5260)
-
-val topAppBarColor = Color(0xFF555555)
-
-val dividerColor = Color(0xFFAAAAAA)
-
-val buttonColor = Color(0xFF04C09E)
-
-val cartPageButtonColor = Color(0xFF04C09E)
